[IMP] web: convert transition tests to hoot task-id: 3705027
